requirements.txt
aiogram
pillow
google.generativeai
python-dotenv